--< Code >-- 보상Item 중복 등장 안 되도록 구현
To Do : 탈부착과 사용 구현 중

+++ MagicSchool/Assets/@Scripts/Contents/Item/ActiveItem.cs
@@ -77,14 +77,9 @@ public class ItemSkill : Equipment
     }
 }
 
-
-
-
-
-
 public class Consumable : ActiveItem
 {
-    public double Value { get; private set; }
+    public float Value { get; private set; }
 
     public Consumable(int templateId) : base(templateId)
     {
